FEAT: Divide search side and add channel search capabilities
